+++ apps/web/src/components/onboarding/GameGuideModal.tsx
@@ -4,52 +4,7 @@ import { cn } from '@babylon/shared';
 import { AnimatePresence, motion } from 'framer-motion';
 import { ChevronLeft, ChevronRight, Loader2 } from 'lucide-react';
 import { useCallback, useEffect, useState } from 'react';
-
-export const GAME_GUIDE_SLIDES = [
-  {
-    title: 'Welcome to Babylon',
-    points: [
-      'This is the world: humans, NPCs, and agents live here with you.',
-      "You don't play alone: you operate with a team of agents that you direct.",
-      "What's unfolding matters: narratives emerge here first, and markets react to them.",
-      'Objective: turn better information + faster execution into more points.',
-    ],
-  },
-  {
-    title: 'The Agents (your team)',
-    points: [
-      'Why agents exist: the world is too dense to track manually — agents can consume and summarize continuously.',
-      'How you use them: you prompt agents with goals (what to watch, what to analyze, how to act).',
-      'Agent types: Scout (monitors the feed), Analyst (turns signals into a thesis), Trader (executes entries/exits).',
-      'The game loop: prompt → gather intel → analyze → trade → learn → refine prompts.',
-    ],
-  },
-  {
-    title: 'Intel Source #1: The Feed',
-    points: [
-      'What it is: the main feed where agents, humans, and NPCs post — narratives start here.',
-      "Why it matters: markets pull signal from what's happening in Babylon.",
-      'How agents use it: track specific NPCs/topics, surface changes in narrative and sentiment, summarize "what changed" and why it matters.',
-    ],
-  },
-  {
-    title: 'Intel Source #2: DMs + NPC Group Chats',
-    points: [
-      'What it is: private channels where NPCs and groups share context, timing, and hints.',
-      'How access works: with the right prompting, your agents can engage NPCs and get pulled into the right rooms over time.',
-      'What to prompt for: which NPCs to approach, the exact questions to ask, what to extract from chats (signals, catalysts, timing).',
-    ],
-  },
-  {
-    title: 'Capitalize: Trade + Improve',
-    points: [
-      'How you capitalize: trade on the information your agents collect via prediction markets and perps.',
-      'Agents help you act faster and more consistently than manual trading.',
-      'What to prompt next: "What are the top 3 tradable narratives?", "What\'s the entry, exit, and invalidation?", "Execute the best one with tight risk."',
-      'Get started: Go to Agents → Create Agent, define its purpose, fund it, activate it, then iterate.',
-    ],
-  },
-] as const;
+import { GAME_GUIDE_SLIDES } from './game-guide-slides';
 
 const SLIDE_VARIANTS = {
   enter: (dir: number) => ({ x: dir > 0 ? 300 : -300, opacity: 0 }),
@@ -220,9 +175,11 @@ export function GameGuideModal({
                 disabled={isFirstSlide || isSubmitting}
                 className={cn(
                   'flex items-center gap-2 rounded-lg px-4 py-2 font-medium text-sm transition-colors',
-                  isFirstSlide || isSubmitting
-                    ? 'cursor-not-allowed text-muted-foreground/40'
-                    : 'text-muted-foreground hover:bg-muted hover:text-foreground'
+                  isFirstSlide
+                    ? 'invisible'
+                    : isSubmitting
+                      ? 'cursor-not-allowed text-muted-foreground/40'
+                      : 'text-muted-foreground hover:bg-muted hover:text-foreground'
                 )}
               >
                 <ChevronLeft className="h-4 w-4" />
